feat: добавить тесты
